Fixed E_TIMEOUT caused by overflow
Whoopsiedaisy I committed a bunch of other stuff.
Oh well.

+++ pingpong/bluetooth.h
@@ -33,7 +33,7 @@ namespace Bluetooth {
    /**
     * We can't use a timer :D /s
     */
-   void listen(unsigned int listenTimeout, unsigned int acceptTimeout = BLUETOOTH_DEFAULT_TIMEOUT);
+   void listen(unsigned long listenTimeout, unsigned long acceptTimeout = BLUETOOTH_DEFAULT_TIMEOUT);
 
 
   /**
